% Auto-generated: do not edit by hand
\name{customReactComponent}

\alias{customReactComponent}

\title{CustomReactComponent component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
customReactComponent(id=NULL, label=NULL, value=NULL, shouldBeTicking=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{value}{Character. The value displayed in the input.}

\item{shouldBeTicking}{Logical. If the counter should start counting on load.}
}
