% Auto-generated: do not edit by hand
\name{customReactRenderComponent}

\alias{customReactRenderComponent}

\title{CustomReactRenderComponent component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
customReactRenderComponent(text=NULL)
}

\arguments{
\item{text}{Character. The ID used to identify this component in Dash callbacks.}
}
