% Auto-generated: do not edit by hand
\name{genericButton}

\alias{genericButton}

\title{GenericButton component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
genericButton(buttonAction=NULL, buttonLabel=NULL)
}

\arguments{
\item{buttonAction}{The ID used to identify this component in Dash callbacks.}

\item{buttonLabel}{Character. }
}
