% Auto-generated: do not edit by hand
\name{stopTickingComponent}

\alias{stopTickingComponent}

\title{StopTickingComponent component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
stopTickingComponent(stopTicking=NULL)
}

\arguments{
\item{stopTicking}{The ID used to identify this component in Dash callbacks.}
}
